# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "my-clothing",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "本地主機"
        },
        {
            "url": "http://dev.laravel-sail.site:8080",
            "description": "expose主機"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "註冊會員",
                "operationId": "f1fde7f4a932e393f6042d1b92e65b0a",
                "requestBody": {
                    "request": "RegisterRequestBody",
                    "description": "註冊會員的請求範例",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "allen"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "allen@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "409": {
                        "description": "email已經被使用過",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "email already exist"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證請求有誤",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/validateError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "登入",
                "operationId": "abc53c091ac89d7a2af0892997d1ee08",
                "requestBody": {
                    "request": "LoginRequestBody",
                    "description": "登入會員的請求範例",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "root@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "user": {
                                        "id": 1,
                                        "name": "allen",
                                        "email": "allen@example.com",
                                        "phone": "0912345678",
                                        "address": null
                                    },
                                    "token": "2|UvEntO7H9iN4SF9eJYPNCxVbBN8V0Nv5TnANWUO9"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "email或password有誤",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Credentials do not match records"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "登出",
                "operationId": "43958ea73917f8e50945f73452c3bff0",
                "responses": {
                    "204": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "用戶資料",
                "operationId": "9cd1cee912e0176094a7ec83cc9b432f",
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/userInfo"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "validateError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The name field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "name": [
                                "The name field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "userInfo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "allen"
                    },
                    "email": {
                        "type": "string",
                        "example": "allen@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0912345678"
                    },
                    "address": {
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "call url /login, paste token value to here.",
                "scheme": "bearer"
            }
        }
    },
...
}